moved kubernetes file handling functions to seperate file for easy importation and use amongst all handlers
